+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Upgrade `rubocop` dependency from `0.15.0` to `0.16.0`
 * Fix broken `--show-linters` flag
 * Ignore `BlockAlignment`, `EndAlignment`, and `IndentationWidth` Rubocop lints
+* Fix bug where `SpaceBeforeScript` linter would incorrectly report lints when
+  the same substring appeared on a line underneath a tag with inline script
 
 ## 0.3.0